Add database replica configuration. (#35)
* Add database replica configuration.

- Updates the bay/settings.php file so that it includes replica access.

This change should allow Drupal Core (and subsequently custom modules) to offload read requests to the replica database. This can be achieved by accessing `\Drupal::services('database.replica')` which will be `NULL` if the replica is not configured. When performing queries against the database we should look to implement them as:

```
$connection = \Drupal::services('database.replica') ?: \Drupal::database();
```

@see https://www.drupal.org/files/issues/2018-09-25/3001216-replica-service-41.patch

* Add 'reader' configuration for drush to target.

* Update the env var for replica support.

- https://github.com/amazeeio/dbaas-mariadb-apb/pull/5/files#diff-e349823b467059e175ddb237223092ecR50

* Update settings.php